refactor: rename importUsersByCSV to correct spelling and remove deprecated files
