+++ src/inngest/types.ts
@@ -25,8 +25,8 @@ import type { CapitolCanaryUpsertAdvocateInngestSchema } from '@/inngest/functio
 import type { CleanupNftMintsEventSchema } from '@/inngest/functions/cleanupNFTMints'
 import type { CleanupPostalCodesInngestEventSchema } from '@/inngest/functions/cleanupPostalCodes'
 import { UpdateDistrictsRankingsCronJobSchema } from '@/inngest/functions/districtsRankings/updateRankings'
-import type { ImportUsersByCSVCoordinatorSchema } from '@/inngest/functions/importlUsersByCSV'
-import { ImportUsersByCSVProcessorSchema } from '@/inngest/functions/importlUsersByCSV/logic'
+import type { ImportUsersByCSVCoordinatorSchema } from '@/inngest/functions/importUsersByCSV'
+import { ImportUsersByCSVProcessorSchema } from '@/inngest/functions/importUsersByCSV/logic'
 import type { InitialSignupUserCommunicationSchema } from '@/inngest/functions/initialSignupUserCommunicationJourney/initialSignupUserCommunicationJourney'
 import type { MonitorBaseEthBalancesInngestEventSchema } from '@/inngest/functions/monitorBaseETHBalances'
 import { SyncSendgridContactsCoordinatorSchema } from '@/inngest/functions/sendgridContactsCronJob'
